+++ CHANGELOG.md
@@ -59,6 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - programs: Fixed bug of allowing limit-swap orders to be updated to accept zero `min_output`.
 - programs: Fixed inconsistent market token balance validation for `GlvDepositOperation`.
 - programs: Fixed incorrect slot used as the publishing slot for a Switchboard feed price. Now the `SbFeed::result_land_slot()` is used instead.
+- programs: Fixed heartbeat validation for Switchboard to be based on `SbFeed::result_ts()`.
 
 ## [0.3.0] - 2025-02-18